Adjust padding
